* vladdu/jinterface_code_cleanup/OTP-12211:
  fix warning: Socket's InputStream not closed
  silence 'potential resource leak' warnings
  don't warn about some unused fields/variables
  handle warnings about potential null pointers
  keep all method parameters final
  renamed method parameters hiding fields
  add 'break' statements to make compiler happy
  remove warnings from OtpMD5
  rename OtpNode.flags to connFlags
  rename OtpNode.Acceptor.port to acceptorPort
  rename field 'self' to 'localNode'
  mark deprecated unused private constructor
  simplify 'if' statements
  remove warnings: @Override, switch fall-through
  make serialVersionUID fields private
  remove unnecessary syntax (casts)
  remove unused variables
  removed unnecessary semicolons, imports, labels
  remove unnecessary null pointer checks
  remove unnecessary Cloneable and Serializable

+++ lib/jinterface/java_src/com/ericsson/otp/erlang/OtpErlangObject.java
@@ -171,14 +171,14 @@ public abstract class OtpErlangObject implements Serializable, Cloneable {
 	    for (j = 0, k = 0;  
 		 j + 4 < b.length;  
 		 j += 4, k += 1, k %= 3) {
-		abc[k] += ((int)b[j+0] & 0xFF) + ((int)b[j+1]<<8 & 0xFF00)
-		    + ((int)b[j+2]<<16 & 0xFF0000) + ((int)b[j+3]<<24);
+        abc[k] += (b[j+0] & 0xFF) + (b[j+1]<<8 & 0xFF00)
+            + (b[j+2]<<16 & 0xFF0000) + (b[j+3]<<24);
 		mix();
 	    }
 	    for (int n = 0, m = 0xFF;
 		 j < b.length;
 		 j++, n += 8, m <<= 8) {
-		abc[k] += (int)b[j]<<n & m;
+        abc[k] += b[j]<<n & m;
 	    }
 	    mix();
 	}
